# services/api/openapi.json
{
  "components": {
    "schemas": {
      "AuditEntry": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "actor_type": {
            "title": "Actor Type",
            "type": "string"
          },
          "at": {
            "title": "At",
            "type": "string"
          },
          "config_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Version"
          },
          "definition_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition Version"
          },
          "evidence": {
            "title": "Evidence"
          },
          "hash": {
            "title": "Hash",
            "type": "string"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "seq",
          "at",
          "actor_type",
          "actor_id",
          "action",
          "subject",
          "evidence",
          "definition_version",
          "config_version",
          "hash"
        ],
        "title": "AuditEntry",
        "type": "object"
      },
      "Body_upload_api_v1_documents_post": {
        "properties": {
          "doc_type": {
            "default": "bill_of_lading",
            "title": "Doc Type",
            "type": "string"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_api_v1_documents_post",
        "type": "object"
      },
...
      "ConfigOut": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "published_at": {
            "title": "Published At",
            "type": "string"
          },
          "published_by": {
            "title": "Published By",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "config",
          "published_by",
          "published_at"
        ],
        "title": "ConfigOut",
        "type": "object"
      },
...
      "DocumentDetail": {
        "properties": {
          "doc_type": {
            "title": "Doc Type",
            "type": "string"
          },
          "extraction": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "run_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Status"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "uploaded_at": {
            "title": "Uploaded At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "doc_type",
          "filename",
          "pages",
          "sha256",
          "uploaded_at",
          "extraction"
        ],
        "title": "DocumentDetail",
        "type": "object"
      },
      "DocumentOut": {
        "properties": {
          "doc_type": {
            "title": "Doc Type",
            "type": "string"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "run_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Status"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "uploaded_at": {
            "title": "Uploaded At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "doc_type",
          "filename",
          "pages",
          "sha256",
          "uploaded_at"
        ],
        "title": "DocumentOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TenantConfig": {
        "additionalProperties": true,
        "description": "Versioned per-tenant knobs read by CEL as `tenant.*`. Unknown keys are allowed on purpose.",
        "properties": {
          "approval_limits": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Approval Limits",
            "type": "object"
          },
          "currency": {
            "default": "USD",
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          }
        },
        "title": "TenantConfig",
        "type": "object"
      },
      "UploadOut": {
        "properties": {
          "doc_type": {
            "title": "Doc Type",
            "type": "string"
          },
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "run_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Status"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "uploaded_at": {
            "title": "Uploaded At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "doc_type",
          "filename",
          "pages",
          "sha256",
          "uploaded_at",
          "duplicate"
        ],
        "title": "UploadOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Verification": {
        "properties": {
          "broken": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Broken",
            "type": "array"
          },
          "entries": {
            "title": "Entries",
            "type": "integer"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid",
          "entries",
          "broken"
        ],
        "title": "Verification",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Nova API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/audit": {
      "get": {
        "operationId": "list_audit_api_v1_audit_get",
        "parameters": [
          {
            "description": "seq to page back from",
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "seq to page back from",
              "title": "Before"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "prefix, e.g. task:<id> or run:<id>",
            "in": "query",
            "name": "subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "prefix, e.g. task:<id> or run:<id>",
              "title": "Subject"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEntry"
                  },
                  "title": "Response List Audit Api V1 Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Audit",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/v1/audit/verify": {
      "get": {
        "operationId": "verify_api_v1_audit_verify_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Verification"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Verify",
        "tags": [
          "audit"
        ]
      }
    },
...
    "/api/v1/documents": {
      "get": {
        "operationId": "list_documents_api_v1_documents_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentOut"
                  },
                  "title": "Response List Documents Api V1 Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Documents",
        "tags": [
          "documents"
        ]
      },
      "post": {
        "operationId": "upload_api_v1_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_v1_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{doc_id}": {
      "get": {
        "operationId": "get_document_api_v1_documents__doc_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{doc_id}/file": {
      "get": {
        "operationId": "get_file_api_v1_documents__doc_id__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get File",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/tenant-config": {
      "get": {
        "operationId": "get_config_api_v1_tenant_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Config",
        "tags": [
          "config"
        ]
      },
      "put": {
        "operationId": "put_config_api_v1_tenant_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Put Config",
        "tags": [
          "config"
        ]
      }
    },
...
  }
}